backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
greenlet==3.2.4

# Redis
redis==5.0.1

# Vector Database
weaviate-client==4.4.0

# LLM Orchestration - LangGraph
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5

# Embeddings - Open Source
sentence-transformers==2.3.1
torch==2.1.2

# Embeddings - Optional Paid APIs
openai==1.10.0
anthropic==0.8.1

# PDF Processing
pdfplumber==0.10.3
pypdf==3.17.4

# Object Storage
minio==7.2.3

# Observability & Monitoring
prometheus-client==0.19.0
langsmith==0.1.142

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSockets
websockets==12.0

# Development
black==23.12.1
ruff==0.1.13
mypy==1.8.0
